lib/ukbitops: Move bitmap.h under lib namespace
This change moves bitmap.h under bitops/ in the default include path,
as is convention for secondary library headers.
Users of bitmap.h have been updated.
