+++ lib/ukbitops/include/uk/bitops/bitmap.h
